+++ remotestate-ts/CHANGES.md
@@ -1,5 +1,10 @@
 ## Version 0.1.1 (in development)
 
+- Refactorings:
+  - Renamed `RemoteState` to `RemoteStateClient`.
+  - Renamed `RemoteStateOptions` to `RemoteStateClientOptions`.
+  - Renamed `createRemoteState` to `createRemoteStateClient`.
+  - Renamed `useRemoteState` to `useRemoteStateClient`.
 
 ## Version 0.1.0